requirements.txt
azure-eventhub-checkpointstoreblob-aio
azure-eventhub
antlr4-python3-runtime
tweepy
